refactor: move session files to a folder
