+++ src/__tests__/use-case/session/db-session-limit-check.spec.ts
@@ -1,6 +1,6 @@
 import { AuthenticationByAccountRepository } from '@/use-case/session/protocols/authentication-by-account-repository'
 import { SessionCountByAuthenticationRepository } from '@/use-case/session/protocols/session-count-by-authentication-repository'
-import { DbSessionLimitCheckByAccount } from '@/use-case/session/session-limit/db-session-limit-check'
+import { DbSessionLimitCheckByAccount } from '@/use-case/session/db-session-limit-check'
 import { makeAuthenticationByAccountStub, mockedAuthentication } from '../stubs/authentications'
 import { makeSessionCountByAuthenticationRepository } from '../stubs/sessions'
