+++ src/use-case/session/db-session-by-account.ts
@@ -1,6 +1,6 @@
 import { Session, SessionByAccount } from '@/entity/session'
-import { AuthenticationByAccountRepository } from '../protocols/authentication-by-account-repository'
-import { SessionListByAuthenticationRepository } from '../protocols/session-repository'
+import { AuthenticationByAccountRepository } from './protocols/authentication-by-account-repository'
+import { SessionListByAuthenticationRepository } from './protocols/session-repository'
 
 export class DbSessionByAccount implements SessionByAccount {
   constructor (
