+++ src/use-case/session/db-session-limit-check.ts
@@ -1,6 +1,6 @@
 import { SessionLimitCheckByAccount } from '@/entity/session'
-import { AuthenticationByAccountRepository } from '../protocols/authentication-by-account-repository'
-import { SessionCountByAuthenticationRepository } from '../protocols/session-count-by-authentication-repository'
+import { AuthenticationByAccountRepository } from './protocols/authentication-by-account-repository'
+import { SessionCountByAuthenticationRepository } from './protocols/session-count-by-authentication-repository'
 
 export class DbSessionLimitCheckByAccount implements SessionLimitCheckByAccount {
   constructor (
